Extract stale-poll + SSE regions into sections/
Two extractions in one commit because sse.js imports checkUpdates and _scheduleCheckUpdates from stale-poll.js and had to land together.

stale-poll.js (358 LOC): the seven module-level stale bindings collapse into a staleState object so dashboard-main.js's _reloadInPlace can reassign staleState.dirtyNodes without hitting the ESM live-binding limit. Moves the companion functions that manipulate the baseline + dirty set (_autoReloadActiveTab, _minimalRoots, updateBaseline, _refreshBaselineFor) plus reloadNode and refreshAll — already imported by other sections from dashboard-main.js, now re-pointed at stale-poll.js.

sse.js (149 LOC): groups _eventSource / reconnect state / configReloadTimer under sseState for the same reason, replaces the bare _startEventStream() trailer call with initSseSideEffects(). Preserves the unreachable _onConfigChanged function with a comment flagging it for a future simplification PR (cut 2 took the same line with its dead window.reloadNode monkey-patch).

dashboard-main.js: 2835 -> 2416 LOC (-419). New section count: 16.

Design: projects/2026-04-23-condash-frontend-extraction/notes/05-p09-cut3.md
